// src/table/state-manager/table-action-types.ts
export enum ActionTableTypesE {
  SET_REMOTE_DATA = "SET_REMOTE_DATA",
  SET_BODY_DATA = "SET_BODY_DATA",
  SET_FILTER_LIMIT = "SET_FILTER_LIMIT",
  SET_FILTER_PAGINATE = "SET_FILTER_PAGINATE",
  SET_SELECT_ALL = "SET_SELECT_ALL",
  SET_SELECTED_ITEMS = "SET_SELECTED_ITEMS",
  SET_FILTER_VALUES = "SET_FILTER_VALUES",
  SET_SORT_CONFIG = "SET_SORT_CONFIG",
  SET_COLUMN_VISIBILITY = "SET_COLUMN_VISIBILITY",
  SET_OPEN_FILTER_BOX = "SET_OPEN_FILTER_BOX",
  SET_IS_COLUMN_MENU_OPEN = "SET_IS_COLUMN_MENU_OPEN",
  SET_FILTER_SEARCH = "SET_FILTER_SEARCH",
  SET_CHECK_BOX = "SET_CHECK_BOX",
  SET_CHECKED_ITEMS = "SET_CHECKED_ITEMS",
  SET_LOADING = "SET_LOADING",
  SET_COLOR='SET_COLOR'
}